# Simplify & optimize style calculation