--- a/js/style/style_declaration.js
+++ b/js/style/style_declaration.js
@@ -3,7 +3,6 @@
 const MapboxGLFunction = require('./style_function');
 const parseColor = require('./parse_color');
 const util = require('../util/util');
-const interpolate = require('../util/interpolate');
 
 module.exports = StyleDeclaration;
 
@@ -23,10 +22,6 @@ function StyleDeclaration(reference, value) {
 
     if (reference.type === 'color') {
         this.calculate = wrapColorCalculate(this.calculate);
-    }
-
-    if (reference.function === 'piecewise-constant' && reference.transition) {
-        this.calculate = wrapTransitionedCalculate(this.calculate);
     }
 
     if (!this.isFeatureConstant && !this.isZoomConstant) {
@@ -53,27 +48,5 @@ function wrapColorCalculate(calculate) {
     return function(globalProperties, featureProperties) {
         const color = calculate(globalProperties, featureProperties);
         return color && parseColor(color);
-    };
-}
-
-// This function is used to smoothly transition between discrete values, such
-// as images and dasharrays.
-function wrapTransitionedCalculate(calculate) {
-    return function(globalProperties, featureProperties, duration) {
-        const z = globalProperties.zoom;
-        const zh = globalProperties.zoomHistory;
-
-        const fromScale = z > zh.lastIntegerZoom ? 2 : 0.5;
-        const from = calculate({zoom: z > zh.lastIntegerZoom ? z - 1 : z + 1}, featureProperties);
-        const to = calculate({zoom: z}, featureProperties);
-
-        const timeFraction = Math.min((Date.now() - zh.lastIntegerZoomTime) / duration, 1);
-        const zoomFraction = Math.abs(z - zh.lastIntegerZoom);
-        const t = interpolate(timeFraction, 1, zoomFraction);
-
-        if (from === undefined || to === undefined)
-            return undefined;
-
-        return { from, fromScale, to, toScale: 1, t };
     };
 }