--- a/js/style/style_declaration.js
+++ b/js/style/style_declaration.js
@@ -1,52 +1,58 @@
 'use strict';
 
-const MapboxGLFunction = require('./style_function');
+const MapboxGLFunction = require('mapbox-gl-function');
 const parseColor = require('./parse_color');
 const util = require('../util/util');
 
-module.exports = StyleDeclaration;
+class StyleDeclaration {
 
-function StyleDeclaration(reference, value) {
-    this.value = util.clone(value);
-    this.isFunction = MapboxGLFunction.isFunctionDefinition(value);
+    constructor(reference, value) {
+        this.value = util.clone(value);
+        this.isFunction = MapboxGLFunction.isFunctionDefinition(value);
 
-    // immutable representation of value. used for comparison
-    this.json = JSON.stringify(this.value);
+        // immutable representation of value. used for comparison
+        this.json = JSON.stringify(this.value);
 
-    const parsedValue = reference.type === 'color' && this.value ? parseColor(this.value) : value;
-    let specDefault = reference.default;
-    if (specDefault && reference.type === 'color') specDefault = parseColor(specDefault);
-    this.calculate = MapboxGLFunction[reference.function || 'piecewise-constant'](parsedValue);
-    this.isFeatureConstant = this.calculate.isFeatureConstant;
-    this.isZoomConstant = this.calculate.isZoomConstant;
+        this.isColor = reference.type === 'color';
 
-    if (reference.type === 'color') {
-        this.calculate = wrapColorCalculate(this.calculate);
-    }
+        const parsedValue = this.isColor && this.value ? parseColor(this.value) : value;
+        let specDefault = reference.default;
+        if (specDefault && reference.type === 'color') specDefault = parseColor(specDefault);
 
-    if (!this.isFeatureConstant && !this.isZoomConstant) {
-        this.stopZoomLevels = [];
-        const interpolationAmountStops = [];
-        const stops = this.value.stops;
-        for (let i = 0; i < this.value.stops.length; i++) {
-            const zoom = stops[i][0].zoom;
-            if (this.stopZoomLevels.indexOf(zoom) < 0) {
-                this.stopZoomLevels.push(zoom);
-                interpolationAmountStops.push([zoom, interpolationAmountStops.length]);
+        this.function = MapboxGLFunction[reference.function || 'piecewise-constant'](parsedValue, specDefault);
+        this.isFeatureConstant = this.function.isFeatureConstant;
+        this.isZoomConstant = this.function.isZoomConstant;
+
+        if (!this.isFeatureConstant && !this.isZoomConstant) {
+            this.stopZoomLevels = [];
+            const interpolationAmountStops = [];
+            for (const stop of this.value.stops) {
+                const zoom = stop[0].zoom;
+                if (this.stopZoomLevels.indexOf(zoom) < 0) {
+                    this.stopZoomLevels.push(zoom);
+                    interpolationAmountStops.push([zoom, interpolationAmountStops.length]);
+                }
             }
-        }
 
-        this.calculateInterpolationT = MapboxGLFunction.interpolated({
-            stops: interpolationAmountStops,
-            base: value.base,
-            colorSpace: value.colorSpace
-        });
+            this.functionInterpolationT = MapboxGLFunction.interpolated({
+                stops: interpolationAmountStops,
+                base: value.base,
+                colorSpace: value.colorSpace
+            });
+        }
+    }
+
+    calculate(globalProperties, featureProperties) {
+        const value = this.function(globalProperties && globalProperties.zoom, featureProperties || {});
+        if (this.isColor && value) {
+            return parseColor(value);
+        }
+        return value;
+    }
+
+    calculateInterpolationT(globalProperties, featureProperties) {
+        return this.functionInterpolationT(globalProperties && globalProperties.zoom, featureProperties || {});
     }
 }
 
-function wrapColorCalculate(calculate) {
-    return function(globalProperties, featureProperties) {
-        const color = calculate(globalProperties, featureProperties);
-        return color && parseColor(color);
-    };
-}
+module.exports = StyleDeclaration;