--- a/js/style/style_transition.js
+++ b/js/style/style_transition.js
@@ -8,29 +8,24 @@ const interpolate = require('../util/interpolate');
  */
 class StyleTransition {
 
-    constructor(reference, declaration, oldTransition, value) {
+    constructor(reference, declaration, oldTransition, options) {
         this.declaration = declaration;
         this.startTime = this.endTime = (new Date()).getTime();
 
-        if (reference.function === 'piecewise-constant' && reference.transition) {
-            this.interp = interpZoomTransitioned;
-        } else {
-            this.interp = interpolate[reference.type];
-        }
-
         this.oldTransition = oldTransition;
-        this.duration = value.duration || 0;
-        this.delay = value.delay || 0;
+        this.duration = options.duration || 0;
+        this.delay = options.delay || 0;
+
+        this.zoomTransitioned = reference.function === 'piecewise-constant' && reference.transition;
+        this.interp = this.zoomTransitioned ? interpZoomTransitioned : interpolate[reference.type];
 
         if (!this.instant()) {
             this.endTime = this.startTime + this.duration + this.delay;
-            this.ease = util.easeCubicInOut;
         }
 
         if (oldTransition && oldTransition.endTime <= this.startTime) {
             // Old transition is done running, so we can
             // delete its reference to its old transition.
-
             delete oldTransition.oldTransition;
         }
     }
@@ -43,19 +38,44 @@ class StyleTransition {
      * Return the value of the transitioning property at zoom level `z` and optional time `t`
      */
     calculate(globalProperties, featureProperties, time) {
-        let value = this.declaration.calculate(globalProperties, featureProperties, this.duration);
-
-        if (this.instant()) return value;
-
-        const t = time || Date.now();
-
-        if (t < this.endTime) {
-            const oldValue = this.oldTransition.calculate(globalProperties, featureProperties, this.startTime);
-            const eased = this.ease((t - this.startTime - this.delay) / this.duration);
-            value = this.interp(oldValue, value, eased);
+        let value;
+        if (this.zoomTransitioned) {
+            value = this._calculateZoomTransitioned(globalProperties, featureProperties);
+        } else {
+            value = this.declaration.calculate(globalProperties, featureProperties);
         }
 
-        return value;
+        if (this.instant())
+            return value;
+
+        time = time || Date.now();
+
+        if (time >= this.endTime)
+            return value;
+
+        const oldValue = this.oldTransition.calculate(globalProperties, featureProperties, this.startTime);
+        const t = util.easeCubicInOut((time - this.startTime - this.delay) / this.duration);
+        return this.interp(oldValue, value, t);
+    }
+
+    // This function is used to smoothly transition between discrete values, such
+    // as images and dasharrays.
+    _calculateZoomTransitioned(globalProperties, featureProperties) {
+        const z = globalProperties.zoom;
+        const zh = globalProperties.zoomHistory;
+
+        const fromScale = z > zh.lastIntegerZoom ? 2 : 0.5;
+        const from = this.declaration.calculate({zoom: z > zh.lastIntegerZoom ? z - 1 : z + 1}, featureProperties);
+        const to = this.declaration.calculate({zoom: z}, featureProperties);
+
+        const timeFraction = Math.min((Date.now() - zh.lastIntegerZoomTime) / this.duration, 1);
+        const zoomFraction = Math.abs(z - zh.lastIntegerZoom);
+        const t = interpolate(timeFraction, 1, zoomFraction);
+
+        if (from === undefined || to === undefined)
+            return undefined;
+
+        return { from, fromScale, to, toScale: 1, t };
     }
 }
 
@@ -64,15 +84,14 @@ module.exports = StyleTransition;
 // This function is used to smoothly transition between discrete values, such
 // as images and dasharrays.
 function interpZoomTransitioned(from, to, t) {
-    if ((from && from.to) === undefined || (to && to.to) === undefined) {
+    if (from === undefined || to === undefined)
         return undefined;
-    } else {
-        return {
-            from: from.to,
-            fromScale: from.toScale,
-            to: to.to,
-            toScale: to.toScale,
-            t: t
-        };
-    }
+
+    return {
+        from: from.to,
+        fromScale: from.toScale,
+        to: to.to,
+        toScale: to.toScale,
+        t: t
+    };
 }