… js/style/style_function.js
… test/js/style/style_declaration.test.js
… test/js/style/style_transition.test.js