A major cleanup of the style recalculation code. 50 less lines of code, much less confusion and also slightly improves performance. All commits are self-contained so review commit by commit.

One noteable architectural change is that `globalProperties` properties like `zoomHistory`, `time` and `duration` that were specific to transition implementation are moved out and live where they belong. So now `globalProperties` are always just `{zoom}`, although I didn't replace it with just `zoom` yet because @lucaswoj anticipates other global properties in future.

benchmark | master 8127e17 | simplify-calculate 55faf5f
--- | --- | ---
**frame-duration** | 6.9 ms, 0% > 16ms  | 6.4 ms, 0% > 16ms 

👀 @lucaswoj @jfirebaugh